requirements.txt
flask
pika
redis
requests
numpy
cupy-cuda12x
xattr


